requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
pytest==5.4.3
psycopg2-binary==2.8.6
